server/requirements.txt
python-dotenv==1.0.1
fastapi==0.111.1
uvicorn==0.30.3
pymongo==4.10.1
pydantic==2.8.2
pyjwt==2.8.0